Rename SeparationClassNamer to ClassDescriptor
