+++ src/main/java/com/github/black0nion/blackonionbot/systems/reload/Reloadable.java
@@ -1,10 +1,10 @@
 package com.github.black0nion.blackonionbot.systems.reload;
 
-import com.github.black0nion.blackonionbot.utils.SeparationClassNamer;
+import com.github.black0nion.blackonionbot.utils.ClassDescriptor;
 
 public interface Reloadable {
 	default String getReloadName() {
-		return SeparationClassNamer.getInstanceClass(this).getSimpleName();
+		return ClassDescriptor.getInstanceClass(this).getSimpleName();
 	}
 
 	void reload();
