+++ src/test/java/com/github/black0nion/blackonionbot/systems/reload/ReloadableTest.java
@@ -1,18 +1,18 @@
 package com.github.black0nion.blackonionbot.systems.reload;
 
-import com.github.black0nion.blackonionbot.utils.SeparationClassNamer;
-import com.github.black0nion.blackonionbot.utils.SeparationClassNamerTest;
+import com.github.black0nion.blackonionbot.utils.ClassDescriptor;
+import com.github.black0nion.blackonionbot.utils.ClassDescriptorTest;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class ReloadableTest extends SeparationClassNamerTest {
+class ReloadableTest extends ClassDescriptorTest {
 
 	@Test
 	void test_getReloadName() {
 		Reloadable reloadable = () -> {};
-		assertEquals(SeparationClassNamer.getInstanceClass(reloadable).getSimpleName(), reloadable.getReloadName());
-		assertEquals(SeparationClassNamer.getInstanceClass(OnlyReloadable.class).getSimpleName(), new OnlyReloadable().getReloadName());
-		assertEquals(SeparationClassNamer.getInstanceClass(ReloadableAndInterface.class).getSimpleName(), new ReloadableAndInterface().getReloadName());
+		assertEquals(ClassDescriptor.getInstanceClass(reloadable).getSimpleName(), reloadable.getReloadName());
+		assertEquals(ClassDescriptor.getInstanceClass(OnlyReloadable.class).getSimpleName(), new OnlyReloadable().getReloadName());
+		assertEquals(ClassDescriptor.getInstanceClass(ReloadableAndInterface.class).getSimpleName(), new ReloadableAndInterface().getReloadName());
 	}
 }
